Add files via upload
I have been working on this project for a while I realized I should be updating when I figure out new functions. I added the settings menu today. Also this is me trying to add the features first and then I'm going to work on the looks of the GUI